requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Login
mysqldbda
Jinja2==2.9.4
MarkupSafe==0.23
WTForms==2.1
Werkzeug==0.11.15
click==6.7
itsdangerous==0.24
wsgiref==0.1.2

# ----------------------
# Production Server
# ----------------------
gunicorn
